move components for evaluations from container folder to component folder
